PYTHON-5947 Split the otel span tests by concern
Three files instead of one, so the upcoming stacked PRs can each carry the
tests for the behavior they add: test_otel.py keeps operations that finish in
a single round trip, test_otel_transactions.py takes the transaction span,
and test_otel_getmore.py takes cursor getMores and change streams.

All 73 tests move unchanged; each file carries its own exporter scaffolding
rather than importing a shared base, since synchro rewrites cross-module
names from test/asynchronous. Counts are unchanged at 196 passed and 3
skipped, splitting 98 / 20 / 28 / 50 across the four otel test modules.
